Create Polynomial Equations
We all know what polynomial equations are and it is one of the common problems given to the beginners when they first start learning C. So, there is a simple program shown below which takes the use of functions in C language and solve the polynomial equation entered by the user provided they also enter the value of the unknown variable x.

For example, the polynomial equation that we use in our program is f(x) = 2x2+3x+1. Now, we ask the user for the value of x. Suppose, x = 2. After putting this value in f(x) = 2*22+ 3*2+1, we get the answer as f(x) = 15.

This program exactly does the same calculation with the help of the fact that we can pass an entire array as our function parameter.