Create Differential-Equations
In this program, we find the value of the derivative of the polynomial equation using the same value of x. For example, we have the quadratic equation f(x) = 2x2+3x+1. The first derivative of this equation would be df(x) = 4x + 3. After the putting x = 2 in the derivative, we get df(x) = 4*2 +3 = 11.